requirements.txt
quart==0.19.9
openai>=1.68.2,<2.0.0
azure-identity==1.15.0
azure-search-documents==11.5.2
azure-storage-blob==12.17.0
azure-cosmos==4.5.0
azure-core==1.34.0

azure-ai-agents==1.0.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b11
gunicorn==20.1.0
uvicorn==0.24.0
hypercorn==0.17.3
aiohttp==3.9.2
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.0
pydantic==2.11.4
pydantic-settings==2.2.1
tiktoken==0.9.0
Jinja2==3.1.6
itsdangerous==2.2.0
Werkzeug==3.1.3
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17